refactor: make autoclean exclude standalone
