+++ pyproject.toml
@@ -75,7 +75,7 @@ dependencies = [
     "autoclean-eeg2source>=0.1.0",
     "autocleaneeg-view",
     # GUI dependencies
-    "PyQt5>=5.15.11",
+    "PyQt6>=6.7.0",
     "mne-qt-browser>=0.3.0",
     # "PySide6>=6.5.0",
     "pymupdf>=1.0.0",
