+++ src/autoclean/tools/__init__.py
@@ -1,5 +1,6 @@
 """GUI and visualization tools."""
 
+from .autoclean_exclude import run_autoclean_exclude
 from .autoclean_review import run_autoclean_review
 
-__all__ = ["run_autoclean_review"]
+__all__ = ["run_autoclean_exclude", "run_autoclean_review"]
